overwrite theme when importing existing theme
It is not possible to delete an imported theme.
(without adapting the saved preferences in the workspace)
Additionally, it is not possible to
import a theme with an existing name again.

This change makes it possible to
import a theme more then once,
overwriting the previous theme associated with that name.

To be able to detect if a theme was already imported,
the preference "importedColorThemeName+index" is added.

Note: it is now possible to import a theme with the same name
as a stock theme. Since the imported themes are read
after the stock themes, the imported theme will get precedence.
The only way to get the "default" stock theme back,
is to remove the entries of the imported theme
in the preferences file.




Signed-off-by: Davy meers <davymeers@hotmail.com>